# Conflicts:
#	.gitignore
#	.idea/php.xml
#	backend-api/app/Models/Boards.php
#	frontend-app/package.json
#	frontend-app/src/App.js
#	frontend-app/src/index.js
#	frontend-app/yarn.lock

+++ backend-api/app/Http/Controllers/CardActivityController.php
@@ -4,57 +4,85 @@
 namespace App\Http\Controllers;
 
 
+use App\Models\BoardCard;
 use App\Models\CardActivity;
 use Illuminate\Http\Request;
+use Illuminate\Http\Response;
 
 class CardActivityController
 {
+    //faltaría asociación
     public function create(Request $request)
     {
         $data = $request->all();
 
-        $activity = new CardActivity([
-            'text' => $data['text'],
-            'card_id' => $data['card_id'],
-        ]);
+        $card = BoardCard::find($data["card_id"]);
 
-        return response()->json($activity);
+        if($card === null) {
+            $code = Response::HTTP_NOT_FOUND;
+            return response()->json(["error" => "Card does not exist"])->setStatusCode($code);
+        }else{
+            $activity = new CardActivity($data);
+
+            $activity->save();
+
+            return response()->json($activity);
+        }
     }
 
     public function findById($id)
     {
         $activity = CardActivity::where('id', $id)->first();
 
+        if($activity === null) {
+            $code = Response::HTTP_NOT_FOUND;
+            return response()->json(["error" => "Activity does not exist"])->setStatusCode($code);
+        }
         return response()->json($activity);
     }
 
-    //REVISAR SI TIENE SENTIDO
-    public function findByCardId(Request $request)
+    public function findByCardId($card_id)
     {
-        $data = $request->all();
+        //faltaría hacer la asociación
 
-        $activities = CardActivity::where('card_id', $data['card_id'])->get();
+        $activity = CardActivity::where('card_id', $card_id)->get();
 
-        return response()->json($activities);
+        if(count($activity) === 0) {
+            $code = Response::HTTP_NOT_FOUND;
+            return response()->json(["error" => "This card has no activities"])->setStatusCode($code);
+        }else{
+            return response()->json($activity);
+        }
     }
 
+    //faltaría hacer la asociación
     public function update(Request $request, $id)
     {
         $activity = CardActivity::where('id', $id)->first();
 
-        $dataFromTheCardActivityToUpdate = $request->all();
+        if($activity === null) {
+            $code = Response::HTTP_NOT_FOUND;
 
-        $activity->update($dataFromTheCardActivityToUpdate);
+            return response()->json(["error" => "Activity does not exist"])->setStatusCode($code);
+        }else{
+            $dataFromTheCardActivityToUpdate = $request->all();
 
-        return response()->json($activity);
+            $activity->update($dataFromTheCardActivityToUpdate);
+
+            return response()->json($activity);
+        }
     }
 
+    //faltaría hacer la asociación
     public function delete($id)
     {
         $activity = CardActivity::where('id', $id)->first();
 
+        if($activity === null) {
+            $code = Response::HTTP_NOT_FOUND;
+            return response()->json(["error" => "Activity does not exist"])->setStatusCode($code);
+        }
         $activity->delete();
-
         return response()->json("Activity deleted!");
     }
 }
